Fix some misc issues from James Molloy's website, add make target for virtualbox

+++ kernel/include/memory/kheap.h
@@ -17,7 +17,7 @@ extern uintptr_t placement_address;
 struct kheap_header {
 	uint32_t magic;
 	bool is_hole;
-	uint32_t size;
+	size_t size;
 };
 
 struct kheap_footer {
@@ -37,12 +37,12 @@ struct kheap_heap {
 extern struct kheap_heap *kheap;
 
 struct kheap_heap *kheap_create(uintptr_t start, uintptr_t end, uintptr_t max, bool supervisor, bool readonly);
-void *kheap_alloc(uint32_t size, bool page_align, struct kheap_heap *heap);
+void *kheap_alloc(size_t size, bool page_align, struct kheap_heap *heap);
 void kheap_free(void *p, struct kheap_heap *heap);
 
-uintptr_t kmalloc_internal(uint32_t size, bool align, uint32_t *phys); // Internal use only.
-uintptr_t kmalloc_a(uint32_t size);  // Page aligned.
-uintptr_t kmalloc_p(uint32_t size, uint32_t *phys); // Returns a physical address.
-uintptr_t kmalloc_ap(uint32_t size, uint32_t *phys); // Page aligned and returns a physical address.
-uintptr_t kmalloc(uint32_t size); // Vanilla (normal).
+uintptr_t kmalloc_internal(size_t size, bool align, uint32_t *phys); // Internal use only.
+uintptr_t kmalloc_a(size_t size);  // Page aligned.
+uintptr_t kmalloc_p(size_t size, uint32_t *phys); // Returns a physical address.
+uintptr_t kmalloc_ap(size_t size, uint32_t *phys); // Page aligned and returns a physical address.
+uintptr_t kmalloc(size_t size); // Vanilla (normal).
 void kfree(void *p);
